+++ pom.xml
@@ -22,6 +22,18 @@ limitations under the License.
     <artifactId>shiro-ehcache3-integration</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>1.7</source>
+                    <target>1.7</target>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
     <name>shiro-ehcache3-integration</name>
 
     <dependencies>
@@ -29,12 +41,19 @@ limitations under the License.
             <groupId>org.ehcache</groupId>
             <artifactId>ehcache</artifactId>
             <version>3.1.2</version>
+            <scope>provided</scope>
         </dependency>
-
         <dependency>
             <groupId>org.apache.shiro</groupId>
             <artifactId>shiro-core</artifactId>
             <version>1.3.2</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.12</version>
+            <scope>test</scope>
         </dependency>
 
     </dependencies>
